fix: keep node:child_process and anyagent node deps out of browser bundle
Continues juspay/kolu#682 (which split @anthropic-ai/claude-agent-sdk
and node:sqlite). Two more integration roots were still being pulled
into the client bundle through kolu-common:

- kolu-git's index.ts re-exports resolve.ts → node:child_process
  (execSync), which surfaced as a dev-mode `Module "node:child_process"
  has been externalized for browser compatibility` error at
  resolve.ts:10.
- anyagent's index.ts re-exports with-db.ts / wal-subscription.ts /
  tail-lines.ts, which pull node:fs / node:path / node:sqlite. Same
  class of bug, currently masked by tree-shaking but a single
  side-effecting import away from breaking.

Both packages now expose a /schemas subpath that exports only zod
schemas and pure types; kolu-common imports from those subpaths.
Server-side consumers continue to import the package roots and
nothing changes for them.
