+++ packages/common/src/index.ts
@@ -4,10 +4,10 @@
 // kolu-git); this module re-exports them and composes aggregate types.
 
 import { z } from "zod";
-import { TaskProgressSchema } from "anyagent";
 // Import from `/schemas` subpaths, not package roots — keeps the
-// client bundle free of `@anthropic-ai/claude-agent-sdk` and `node:sqlite`
-// (see juspay/kolu#682).
+// client bundle free of `@anthropic-ai/claude-agent-sdk`, `node:sqlite`,
+// `node:child_process`, etc. (see juspay/kolu#682).
+import { TaskProgressSchema } from "anyagent/schemas";
 import { ClaudeCodeInfoSchema } from "kolu-claude-code/schemas";
 import { CodexInfoSchema } from "kolu-codex/schemas";
 import { OpenCodeInfoSchema } from "kolu-opencode/schemas";
@@ -29,7 +29,7 @@ import {
   FsListDirOutputSchema,
   FsReadFileInputSchema,
   FsReadFileOutputSchema,
-} from "kolu-git";
+} from "kolu-git/schemas";
 
 // Re-export integration schemas so consumers import from kolu-common only.
 export {
@@ -68,7 +68,7 @@ export type {
   GitStatusOutput,
   GitDiffOutput,
   FsListDirOutput,
-} from "kolu-git";
+} from "kolu-git/schemas";
 
 // --- Zod schemas ---
