+++ packages/integrations/anyagent/src/index.ts
@@ -24,24 +24,8 @@ export {
 
 export { readTailLines, type TailReadConfig } from "./tail-lines.ts";
 
-import { z } from "zod";
-
-/** Task/todo progress — total items and completed count.
- *  Used by both Claude Code (from TaskCreate/TaskUpdate tool calls)
- *  and OpenCode (from the `todo` SQLite table). */
-export const TaskProgressSchema = z.object({
-  total: z.number(),
-  completed: z.number(),
-});
-
-export type TaskProgress = z.infer<typeof TaskProgressSchema>;
-
-/** Logger interface accepted by integration library functions.
- *  Structurally compatible with pino child loggers — the server
- *  creates a `log.child(...)` and passes it through. */
-export type Logger = {
-  debug: (obj: Record<string, unknown>, msg: string) => void;
-  info: (obj: Record<string, unknown>, msg: string) => void;
-  warn: (obj: Record<string, unknown>, msg: string) => void;
-  error: (obj: Record<string, unknown>, msg: string) => void;
-};
+export {
+  TaskProgressSchema,
+  type TaskProgress,
+  type Logger,
+} from "./schemas.ts";
